Move from gemini logic to using the openai sdk
