+++ chatapi/package.json
@@ -32,7 +32,6 @@
     "@types/express": "^4.17.17",
     "@types/ws": "^8.5.6",
     "@types/textract": "^2.4.5",
-    "@google/generative-ai": "^0.3.0",
     "cors": "^2.8.5",
     "dotenv": "^16.1.4",
     "express": "^4.18.2",
