--- chatapi/src/models/ai-providers.model.ts
@@ -1,24 +1,0 @@
-type ProviderName = 'openai' | 'perplexity' | 'gemini';
-
-export interface AIProvider {
-  name: ProviderName;
-  model?: string;
-}
-
-interface Assistant {
-  name: string;
-  instructions: string;
-}
-
-interface Providers {
-  openai?: string;
-  perplexity?: string;
-  deepseek?: string;
-  gemini?: string;
-}
-
-export interface ModelsDocument {
-  models: Providers;
-  keys: Providers;
-  assistant?: Assistant;
-}
