--- chatapi/src/models/chat-item.model.ts
@@ -1,4 +1,0 @@
-export interface ChatItem {
-  query: string;
-  response: string;
-}
