--- chatapi/src/models/chat-message.model.ts
@@ -1,9 +1,0 @@
-export interface ChatMessage {
-  role: 'user' | 'assistant';
-  content: string;
-}
-
-export interface GeminiMessage {
-  role: 'user' | 'model';
-  parts: any[];
-}
