+++ chatapi/src/services/chat.service.ts
@@ -1,11 +1,10 @@
 import { DocumentInsertResponse } from 'nano';
 
 import { chatDB } from '../config/nano.config';
-import { aiChat } from '../utils/chat.utils';
 import { retrieveChatHistory } from '../utils/db.utils';
+import { aiChat } from '../utils/chat.utils';
 import { handleChatError } from '../utils/chat-error.utils';
-import { AIProvider } from '../models/ai-providers.model';
-import { ChatMessage } from '../models/chat-message.model';
+import { AIProvider, ChatMessage } from '../models/chat.model';
 
 /**
  * Create a chat conversation & save in couchdb
@@ -24,10 +23,6 @@ export async function chat(data: any, stream?: boolean, callback?: (response: st
 
   if (!content || typeof content !== 'string') {
     throw new Error('"data.content" is a required non-empty string field');
-  }
-
-  if (stream && aiProvider.name === 'gemini') {
-    throw new Error('Streaming not supported on Gemini');
   }
 
   if (dbData._id) {
