+++ chatapi/src/utils/chat-helpers.utils.ts
@@ -1,7 +1,6 @@
 import { keys } from '../config/ai-providers.config';
 import { models } from '../config/ai-providers.config';
-import { AIProvider } from '../models/ai-providers.model';
-import { ChatMessage, GeminiMessage } from '../models/chat-message.model';
+import { AIProvider, ChatMessage } from '../models/chat.model';
 import { Attachment } from '../models/db-doc.model';
 import { fetchFileFromCouchDB } from './db.utils';
 import {
@@ -14,42 +13,6 @@ import {
   createAndHandleRunWithStreaming,
 } from './chat-assistant.utils';
 import { extractTextFromDocument } from './text-extraction.utils';
-
-/**
- * Uses geminis's multimodal endpoint to generate chat completions
- * @param messages - Array of chat messages
- * @param model - Gemini model to use for completions
- * @returns Completion text
- */
-async function handleGemini(
-  messages: ChatMessage[],
-  model: string
-): Promise<string> {
-  const geminiModel = keys.gemini.getGenerativeModel({ model });
-
-  const msg = messages[messages.length - 1].content;
-
-  const geminiMessages: GeminiMessage[] = messages.map((message) => ({
-    'role': message.role === 'assistant' ? 'model' : message.role,
-    'parts': [ { 'text': message.content } ],
-  }));
-
-  geminiMessages.pop();
-
-  const chat = geminiModel.startChat({
-    'history': geminiMessages,
-    'generationConfig': {
-      'maxOutputTokens': 100,
-    },
-  });
-
-  const result = await chat.sendMessage(msg);
-  const response = await result.response;
-  const completionText = response.text();
-
-  return completionText;
-}
-
 
 /**
  * Uses openai's completions endpoint to generate chat completions with streaming enabled
@@ -85,30 +48,24 @@ export async function aiChatStream(
     }
   }
 
-  if (aiProvider.name === 'gemini') {
-    return handleGemini(messages, model);
-  } else if ('chat' in provider.ai) {
-    const completion = await provider.ai.chat.completions.create({
-      model,
-      messages,
-      'stream': true,
-    });
+  const completion = await provider.ai.chat.completions.create({
+    model,
+    messages,
+    'stream': true,
+  });
 
-    let completionText = '';
-    for await (const chunk of completion) {
-      if (chunk.choices && chunk.choices.length > 0) {
-        const response = chunk.choices[0].delta?.content || '';
-        completionText += response;
-        if (callback) {
-          callback(response);
-        }
+  let completionText = '';
+  for await (const chunk of completion) {
+    if (chunk.choices && chunk.choices.length > 0) {
+      const response = chunk.choices[0].delta?.content || '';
+      completionText += response;
+      if (callback) {
+        callback(response);
       }
     }
-
-    return completionText;
-  } else {
-    throw new Error('Provider does not support chat completions');
   }
+
+  return completionText;
 }
 
 
@@ -159,21 +116,15 @@ export async function aiChatNonStream(
     }
   }
 
-  if (aiProvider.name === 'gemini') {
-    return handleGemini(messages, model);
-  } else if ('chat' in provider.ai) {
-    const completion = await provider.ai.chat.completions.create({
-      model,
-      messages,
-    });
+  const completion = await provider.ai.chat.completions.create({
+    model,
+    messages,
+  });
 
-    const completionText = completion.choices[0]?.message?.content;
-    if (!completionText) {
-      throw new Error('Unexpected API response');
-    }
-
-    return completionText;
-  } else {
-    throw new Error('Provider does not support chat completions');
+  const completionText = completion.choices[0]?.message?.content;
+  if (!completionText) {
+    throw new Error('Unexpected API response');
   }
+
+  return completionText;
 }
